requirements.txt
aiogram==3.15.0
yt-dlp==2024.12.23
requests==2.31.0
python-dotenv==1.0.0
deep-translator==1.11.4
aiohttp==3.9.1
